agent-demo/pyproject.toml
[project]
name = "agent-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.84.0",
    "httpx>=0.28.1",
    "openai>=2.28.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
]
